// FE_Manage_Event/src/app/models/user/user.model.ts
export class User {
  id?: string;
  email?: string;
  userName?: string;
  photoURL?: string;
  password?:string;
  role?: string;
}

// export class User {
//     id: number;
//     username: string;
//     birthday: number;
//     email: string;
//     phone: number;
//     major: string;
// }

// export class User {
//     id?: number;
//     userName?: string;
//     password?: string;
//     fullName?: string;
//     email?: string;
//     rolesList?: [{
//         idRole: number;
//         nameRole: string;
//     }];
// }